Add async background task to run TeeBench
I had several problems:
- When putting the channel sender into the ServerState struct it
just didn't work. I could clone the channel, wrap it in an Arc...
For that I discovered the section "Debugging handler type errors"
in axum's docs:
https://docs.rs/axum/0.5.17/axum/handler/index.html#debugging-handler-type-errors

I added the macro they recommend there. It told me that the
MutexGuard around the ServerState is not `Send`. The problem only
appeared when I actually used the channel inside the MutexGuard.
- They seem to have released axum 0.6. I'm still on 0.5.17.
I looked at the 0.6 docs and they introduce a `State`, which seems
to be the new recommended way share state between handlers. Maybe
I should switch to the newer version.