Improve sort order
The derived compare function on the arguments sorts the args
according to the fields, in the order the fields appear in the
source code. So with Platform and Dataset on top, those are the
first criteria for sorting. Then algorithm and now selectivity,
data skew and threads (instead of seal chunk size)