Add missing axes' labels
Closes #53

Also:
Add the proper name of the commit instead of "Latest Operator" to
the chart

Possible Bug: I restarted the `./run.sh` script and immediately
reloaded the webpage. It crashed with a Compile job in the job
results view. How did a compile job end up in there? It might had
something to do with me reenabling session storage for the
`FinishedJobState` and not clearing the session storage in the
browser. Still weird. Went away with clearing it and reloading.